Use c_double instead of c_float for array outputs
This matches changes made for bmi-fortran v1.0.

+++ bmi/bmi_interoperability.f90
@@ -351,7 +351,7 @@ contains
     integer (c_int), intent(in), value :: model_index
     integer (c_int), intent(in), value :: grid_id
     integer (c_int), intent(in), value :: n
-    real (c_float), intent(out) :: grid_spacing(n)
+    real (c_double), intent(out) :: grid_spacing(n)
     integer (c_int) :: status
 
     status = model_array(model_index)%get_grid_spacing(grid_id, grid_spacing)
@@ -365,7 +365,7 @@ contains
     integer (c_int), intent(in), value :: model_index
     integer (c_int), intent(in), value :: grid_id
     integer (c_int), intent(in), value :: n
-    real (c_float), intent(out) :: grid_origin(n)
+    real (c_double), intent(out) :: grid_origin(n)
     integer (c_int) :: status
 
     status = model_array(model_index)%get_grid_origin(grid_id, grid_origin)
@@ -379,7 +379,7 @@ contains
     integer (c_int), intent(in), value :: model_index
     integer (c_int), intent(in), value :: grid_id
     integer (c_int), intent(in), value :: n
-    real (c_float), intent(out) :: grid_x(n)
+    real (c_double), intent(out) :: grid_x(n)
     integer (c_int) :: status
 
     status = model_array(model_index)%get_grid_x(grid_id, grid_x)
@@ -393,7 +393,7 @@ contains
     integer (c_int), intent(in), value :: model_index
     integer (c_int), intent(in), value :: grid_id
     integer (c_int), intent(in), value :: n
-    real (c_float), intent(out) :: grid_y(n)
+    real (c_double), intent(out) :: grid_y(n)
     integer (c_int) :: status
 
     status = model_array(model_index)%get_grid_y(grid_id, grid_y)
@@ -407,7 +407,7 @@ contains
     integer (c_int), intent(in), value :: model_index
     integer (c_int), intent(in), value :: grid_id
     integer (c_int), intent(in), value :: n
-    real (c_float), intent(out) :: grid_z(n)
+    real (c_double), intent(out) :: grid_z(n)
     integer (c_int) :: status
 
     status = model_array(model_index)%get_grid_z(grid_id, grid_z)
